+++ helpers/saveOxygenMasterOutputs.m
@@ -32,12 +32,16 @@ Mean_OxySurge_TraceZ = OutputData.Mean_OxySurge_TraceZ;
 OxySurge_Map = OutputData.OxySurge_Map;
 SurgeCandidateRunQC=OutputData.SurgeCandidateRunQC;
 SurgeGapReview=OutputData.SurgeGapReview;
+SurgeTrackingEdges=OutputData.SurgeTrackingEdges;
+SurgeContactFrames=OutputData.SurgeContactFrames;
 writetable(SurgeCandidateRunQC,fullfile(OutputFolders.OxySurgesPath,'SurgeCandidateRunQC.csv'));
 writetable(SurgeGapReview,fullfile(OutputFolders.OxySurgesPath,'SurgeGapReview.csv'));
+writetable(SurgeTrackingEdges,fullfile(OutputFolders.OxySurgesPath,'SurgeTrackingEdges.csv'));
+writetable(SurgeContactFrames,fullfile(OutputFolders.OxySurgesPath,'SurgeContactFrames.csv'));
 
 save(fullfile(OutputFolders.OxySurgesPath,['OxygenSurges',OutputData.DatafileID,'.mat']), ...
     'Table_OxygenSurges_Out','Table_OxygenSurgeEvents_Out','OxySurgeArea_all','Mean_ROI_TraceZ', ...
-    'Mean_OxySurge_TraceZ','OxySurge_Map','AnalysisInfo','SurgeCandidateRunQC','SurgeGapReview');
+    'Mean_OxySurge_TraceZ','OxySurge_Map','AnalysisInfo','SurgeCandidateRunQC','SurgeGapReview','SurgeTrackingEdges','SurgeContactFrames');
 
 save(fullfile(OutputFolders.ManualCurOxySinksPath,['ManualCuration',OutputData.DatafileID,'.mat']), ...
     'Table_OxygenSinks_Out','Table_OxygenSinkEvents_Out','OxySinkArea_all','Mean_OxySink_TraceZ', ...